vobiz-voice-ai/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LiveKit Agents (Voice AI Platform) - UPDATED FOR MCP
livekit-agents>=1.0.0
livekit-plugins-google>=0.1.0

# MCP (Model Context Protocol) - CRITICAL FOR TOOLS
mcp>=1.0.0

# Google Gemini AI - UPDATED FOR GEMINI 2.5 LIVE
google-genai>=0.3.0
google-generativeai>=0.8.0

# Telephony & WebRTC
aiohttp==3.9.1
websockets==12.0
aiortc==1.7.0

# Database (Supabase)
supabase==2.0.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# n8n Integration
httpx==0.26.0
requests==2.31.0

# Utilities
python-multipart==0.0.6
structlog==24.1.0
prometheus-client==0.19.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
ruff==0.1.11
